Move InMemoryMerkle stuff into internal package
