+++ integration/log.go
@@ -27,9 +27,9 @@ import (
 	"github.com/golang/glog"
 	"github.com/google/trillian"
 	"github.com/google/trillian/client/backoff"
+	"github.com/google/trillian/internal/merkle/memory"
 	"github.com/google/trillian/merkle/compact"
 	"github.com/google/trillian/merkle/logverifier"
-	"github.com/google/trillian/merkle/memory"
 	"github.com/google/trillian/merkle/rfc6962"
 	"github.com/google/trillian/types"
 )
