+++ internal/merkle/memory/memory_merkle_tree.go
